--- example/.grpc-gen.json
@@ -1,10 +1,0 @@
-{
-	"srcs_dir": "./src",
-	"srcs": [
-		"test.proto",
-		"deep/deep.proto"
-	],
-	"node_out": "./dist/node",
-	"web_out": "./dist/web",
-	"ngx_out": "./dist/ngx"
-}
