+++ example/.grpc-gen.yml
@@ -1,0 +1,8 @@
+include_well_known_types: false
+srcs_dir: ./src
+srcs:
+  - test.proto
+  - deep/deep.proto
+output:
+  js: ./dist/node
+  grpc-web: ./dist/web
