+++ package.json
@@ -1,8 +1,8 @@
 {
   "name": "grpc-gen",
-  "version": "0.5.2",
+  "version": "1.0.0-0",
   "bin": "src/cli.js",
-  "description": "Simple CLI for generating stubs for gRPC Node. and grpc-web-client",
+  "description": "Simple CLI for running the protoc compiler via configuration file",
   "main": "src/main",
   "scripts": {},
   "author": "Ezekiel Warren",
@@ -12,18 +12,12 @@
   },
   "license": "ISC",
   "dependencies": {
-    "@angular/compiler": "^5.2.9",
-    "@angular/compiler-cli": "^5.2.9",
-    "@types/google-protobuf": "^3.2.7",
-    "colors": "^1.2.1",
-    "fs-extra": "^5.0.0",
-    "google-protobuf": "^3.5.0",
-    "grpc": "^1.10.1",
-    "grpc-tools": "^1.6.6",
+    "cli-progress": "^1.8.0",
+    "colors": "^1.2.5",
+    "fs-extra": "^6.0.1",
     "js-yaml": "^3.11.0",
-    "protoc-gen-ngx": "^0.1.1",
     "tmp": "0.0.33",
-    "ts-protoc-gen": "^0.5.2",
-    "typescript": "^2.8.1"
+    "unzip": "^0.1.11",
+    "yargs": "^11.0.0"
   }
 }
